# specs/001-digital-library-forum/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Library Forum Backend API
  version: 0.1.0
  description: REST contract for the digital library, academic forum, study groups, moderation, administration, statistics, and system configuration backend.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      responses:
        '200': { description: Current profile, content: { application/json: { schema: { $ref: '#/components/schemas/UserResponse' } } } }
    patch:
      tags: [Users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200': { description: Profile updated, content: { application/json: { schema: { $ref: '#/components/schemas/UserResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
  /users/me/password:
    patch:
      tags: [Users]
      summary: Change current user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204': { description: Password changed }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /documents:
    get:
      tags: [LibraryDocuments]
      security: []
      summary: Search public or accessible documents
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: categoryId, in: query, schema: { type: string, format: uuid } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: Documents returned, content: { application/json: { schema: { $ref: '#/components/schemas/PagedDocumentResponse' } } } }
  /documents/{documentId}:
    get:
      tags: [LibraryDocuments]
      security: []
      summary: View public or accessible document details
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200': { description: Document details, content: { application/json: { schema: { $ref: '#/components/schemas/DocumentResponse' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /documents/{documentId}/read:
    get:
      tags: [LibraryDocuments]
      summary: Read document online
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200': { description: Read URL or stream metadata }
        '403': { $ref: '#/components/responses/Forbidden' }
  /documents/{documentId}/download:
    get:
      tags: [LibraryDocuments]
      summary: Download document file
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200': { description: File stream }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /documents/{documentId}/favorite:
    post:
      tags: [LibraryDocuments]
      summary: Save document as favorite
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '201': { description: Favorite created }
    delete:
      tags: [LibraryDocuments]
      summary: Remove document from favorites
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '204': { description: Favorite removed }
  /documents/{documentId}/ratings:
    post:
      tags: [LibraryDocuments]
      summary: Rate a document
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateDocumentRequest'
      responses:
        '201': { description: Rating saved }
        '400': { $ref: '#/components/responses/ValidationError' }
  /documents/{documentId}/reports:
    post:
      tags: [LibraryDocuments]
      summary: Report a document violation
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201': { description: Report created }
...
  /lecturer/documents/{documentId}:
    patch:
      tags: [LecturerDocuments]
      summary: Update own lecturer document
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200': { description: Document updated }
    delete:
      tags: [LecturerDocuments]
      summary: Delete own lecturer document
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '204': { description: Document deleted }
...
  /study-groups:
    get:
      tags: [StudyGroups]
      summary: List joined or discoverable study groups
      responses:
        '200': { description: Study groups }
    post:
      tags: [StudyGroups]
      summary: Create a study group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudyGroupRequest'
      responses:
        '201': { description: Group created }
  /study-groups/{groupId}/join:
    post:
      tags: [StudyGroups]
      summary: Request or join a study group
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '201': { description: Membership request created or membership approved }
  /study-groups/{groupId}/members/{userId}/approve:
    post:
      tags: [StudyGroups]
      summary: Approve group membership request
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200': { description: Member approved }
  /study-groups/{groupId}/documents:
    get:
      tags: [StudyGroups]
      summary: List group documents
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200': { description: Group documents }
    post:
      tags: [StudyGroups]
      summary: Share a document with group
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '201': { description: Document shared }
  /study-groups/{groupId}/posts:
    post:
      tags: [StudyGroups]
      summary: Create group post
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '201': { description: Group post created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DocumentId: { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
...
    GroupId: { name: groupId, in: path, required: true, schema: { type: string, format: uuid } }
    UserId: { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Role or ownership permission denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        status: { type: string }
        roles: { type: array, items: { type: string } }
    UpdateProfileRequest:
      type: object
      properties:
        fullName: { type: string, minLength: 1, maxLength: 120 }
        avatarUrl: { type: string }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 8 }
    DocumentResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { type: string }
        ownerId: { type: string, format: uuid }
    PagedDocumentResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/DocumentResponse' }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    RateDocumentRequest:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, maxLength: 1000 }
    CreateReportRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
...
    CreateStudyGroupRequest:
      type: object
      required: [name, visibility]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string }
        visibility: { type: string, enum: [PUBLIC, REQUEST_TO_JOIN, PRIVATE] }
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        errors:
          type: array
          items: { type: string }